backend/requirements.txt
# DocuMind AI - Backend Dependencies

# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing (for next phase)
PyMuPDF==1.23.8
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.2.0

# NLP & Analysis (for next phase)
spacy==3.7.2
vaderSentiment==3.3.2
sumy==0.11.0
nltk==3.8.1

# Utilities
httpx==0.26.0
